requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
psycopg2-binary
python-dotenv
aiohttp